address comments from https://docs.google.com/document/d/140kknAtq2MWIWPUMIH4odoHtO0PhnHcP3EjMuAVhwU0/edit
